Move apiEndpointModel to path /model
